requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
fastapi==0.136.0
starlette==0.49.1
uvicorn[standard]==0.34.1
gunicorn==23.0.0
pydantic==2.11.7
pydantic-settings==2.10.1

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.36
alembic==1.18.4
psycopg2-binary==2.9.10

# ── Security ──────────────────────────────────────────────────────────────────
cryptography==46.0.7
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
PyJWT==2.12.1
slowapi==0.1.9

# ── HTTP & Parsing ───────────────────────────────────────────────────────────
httpx==0.28.1
aiofiles==24.1.0
beautifulsoup4==4.13.4
lxml==5.3.2
sse-starlette==2.4.1

# ── Redis ─────────────────────────────────────────────────────────────────────
redis==5.2.1

# ── Monitoring ────────────────────────────────────────────────────────────────
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.0.2
sentry-sdk[fastapi]==2.19.2

# ── Distributed Tracing (optional) ──────────────────────────────────────────
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0

# ── Browser Engine ───────────────────────────────────────────────────────────
playwright==1.58.0

# ── Email validation ─────────────────────────────────────────────────────────
email-validator==2.3.0

# ── MCP (Model Context Protocol) ────────────────────────────────────────────
mcp==1.27.0

# ── CLI ──────────────────────────────────────────────────────────────────────
click==8.1.8